+++ CHANGELOG.md
@@ -20,6 +20,8 @@ We use *breaking* word for marking changes that are not backward compatible (rel
 
 - [#2658](https://github.com/thanos-io/thanos/pull/2658) Upgrade to Prometheus [@f4dd45609a05](https://github.com/prometheus/prometheus/commit/f4dd45609a05) which is after v2.18.1.
     - TSDB now does memory-mapping of Head chunks and reduces memory usage.
+- [#2667](https://github.com/thanos-io/thanos/pull/2667) Store: removed support to the legacy `index.cache.json`. The hidden flag `--store.disable-index-header` was removed.
+- [#2667](https://github.com/thanos-io/thanos/pull/2667) Compact: the deprecated flag `--index.generate-missing-cache-file` and the metric `thanos_compact_generated_index_total` were removed.
 
 ## [v0.13.0](https://github.com/thanos-io/thanos/releases) - IN PROGRESS